+++ NEWS.md
@@ -1,6 +1,7 @@
 # pmatch 0.1.2
 
- * 
+ * Function cases_expr generates an expression (a call object) based on pattern rules.
+ * Exporting test_pattern and test_pattern_ to test one pattern against a value.
 
 # pmatch 0.1.1